Quick Updates but big ones
Enemy will look for where player was last seen, then return to patrolling. A little bugged but will fix post haste. Enemy will chase player when player in range, but return to patrol when player is no longer in range/visible. Player and Enemy now have the functionality to take damage, but this needs to be applied.